requirements.txt
streamlit
moviepy
SpeechRecognition
pydub
google-generativeai
numpy
librosa
edge-tts
Pillow
python-dotenv
soundfile